Allow access to Truncate and NormalizeSInt through AdvancedNumeric
